+++ model.h
@@ -4,14 +4,14 @@
 #include "Ray.h"
 
 struct hit_record {
-    float t;
+    double t;
     vec3 p;
     vec3 normal;
 };
 
 class model{
     public:
-        virtual bool hit(const Ray& r, float t_min, float t_max, hit_record& rec) const = 0;
+        virtual bool hit(const Ray& r, double t_min, double t_max, hit_record& rec) const = 0;
 };
 
 #endif
